Seperate out sections in sample app
