+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -28,6 +28,7 @@ object Dependencies {
         const val appcompat = "androidx.appcompat:appcompat:1.2.0-rc01"
         const val coreKtx = "androidx.core:core-ktx:1.5.0-alpha03"
         const val composeActivity = "androidx.activity:activity-compose:1.3.0-alpha06"
+        const val composeNav = "androidx.navigation:navigation-compose:1.0.0-alpha10"
 
         object Compose {
             const val version = "1.0.0-beta04"
