+++ core/src/main/java/com/vanpra/composematerialdialogs/MaterialDialogButtons.kt
@@ -4,6 +4,8 @@ import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
@@ -13,7 +15,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.layout.Layout
 import androidx.compose.ui.layout.layoutId
+import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.unit.dp
 import java.util.Locale
@@ -23,6 +27,72 @@ internal enum class MaterialDialogButtonTypes {
     Positive,
     Negative,
     Accessibility
+}
+
+/**
+ *  Adds buttons to the bottom of the dialog
+ * @param content the buttons which should be displayed in the dialog.
+ * See [MaterialDialogButtons] for more information about the content
+ */
+@Composable
+fun MaterialDialog.buttons(content: @Composable MaterialDialogButtons.() -> Unit) {
+    val interButtonPadding = with(LocalDensity.current) { 12.dp.toPx().toInt() }
+    val defaultBoxHeight = with(LocalDensity.current) { 36.dp.toPx().toInt() }
+    val accessibilityPadding = with(LocalDensity.current) { 12.dp.toPx().toInt() }
+
+    Box(
+        Modifier
+            .fillMaxWidth()
+            .padding(top = 8.dp, bottom = 8.dp, end = 8.dp)
+            .layoutId("buttons")
+    ) {
+        Layout(
+            { content(buttons) }, Modifier,
+            { measurables, constraints ->
+                val placeables = measurables.map {
+                    (it.layoutId as MaterialDialogButtonTypes) to it.measure(constraints)
+                }
+                val totalWidth = placeables.map { it.second.width }.sum()
+                val column = totalWidth > 0.8 * constraints.maxWidth
+
+                val height =
+                    if (column) {
+                        val buttonHeight = placeables.map { it.second.height }.sum()
+                        val heightPadding = (placeables.size - 1) * interButtonPadding
+                        buttonHeight + heightPadding
+                    } else {
+                        defaultBoxHeight
+                    }
+
+                layout(constraints.maxWidth, height) {
+                    var currX = constraints.maxWidth
+                    var currY = 0
+
+                    val posButtons = placeables.buttons(MaterialDialogButtonTypes.Positive)
+                    val negButtons = placeables.buttons(MaterialDialogButtonTypes.Negative)
+                    val textButtons = placeables.buttons(MaterialDialogButtonTypes.Text)
+                    val accButtons = placeables.buttons(MaterialDialogButtonTypes.Accessibility)
+
+                    val buttonInOrder = posButtons + textButtons + negButtons
+                    buttonInOrder.forEach { button ->
+                        if (column) {
+                            button.place(currX - button.width, currY)
+                            currY += button.height + interButtonPadding
+                        } else {
+                            currX -= button.width
+                            button.place(currX, 0)
+                        }
+                    }
+
+                    if (accButtons.isNotEmpty()) {
+                        /* There can only be one accessibility button so take first */
+                        val button = accButtons[0]
+                        button.place(accessibilityPadding, height - button.height)
+                    }
+                }
+            }
+        )
+    }
 }
 
 /**
